Add Ascetic modifier (#1312)
